requirements.txt
fastapi
uvicorn
streamlit
pandas
requests
python-dotenv
pydantic